# nestjs-project/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/videos/{publicId}": {
      "get": {
        "description": "Returns the current status and metadata of a video. Only the owning channel can access this endpoint.",
        "operationId": "VideosController_getVideo",
        "parameters": [
          {
            "name": "publicId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Video found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VideoDetailResponse"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated user does not own the video",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Video not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Get video status",
        "tags": [
          "videos"
        ]
      },
      "patch": {
        "description": "Edits title, description, category and visibility, replaces the custom thumbnail and publishes or unpublishes the video in a single multipart call. Only the owning channel can access this endpoint.",
        "operationId": "VideosController_updateVideo",
        "parameters": [
          {
            "name": "publicId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 200
                  },
                  "description": {
                    "type": "string",
                    "nullable": true
                  },
                  "category": {
                    "type": "string",
                    "enum": [
                      "Música",
                      "Jogos",
                      "Educação",
                      "Entretenimento",
                      "Notícias",
                      "Esportes",
                      "Tecnologia",
                      "Outros"
                    ]
                  },
                  "visibility": {
                    "type": "string",
                    "enum": [
                      "public",
                      "unlisted"
                    ]
                  },
                  "published": {
                    "type": "boolean"
                  },
                  "thumbnail": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Video updated — same shape as GET /videos/{publicId}",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VideoDetailResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed or unsupported thumbnail type",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated user does not own the video",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Video not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "Video is still processing or failed, so it cannot be published",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Update a video",
        "tags": [
          "videos"
        ]
      }
    },
    "/videos/{publicId}/public": {
      "get": {
        "description": "Returns the display metadata of a published video plus two 6 h presigned storage URLs (inline playback and attachment download). Accessible without authentication. A valid bearer token is optional and only identifies the channel owner, who may also read their own drafts.",
        "operationId": "VideosController_getPublicVideo",
        "parameters": [
          {
            "name": "publicId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Video found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicVideoDetailResponse"
                }
              }
            }
          },
          "404": {
            "description": "Video not found, or a draft requested by someone else",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "Video is not ready for viewing",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Get a published video for public viewing",
        "tags": [
          "videos"
        ]
      }
    },
    "/videos/{publicId}/view": {
      "post": {
        "description": "Increments the view counter. Called by the player once per mount, after enough effective playback has accumulated. Accessible without authentication and takes no body. Rate limited to 30 requests per 60 s per IP.",
        "operationId": "VideosController_registerView",
        "parameters": [
          {
            "name": "publicId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "View registered"
          },
          "404": {
            "description": "Video not found, or a draft requested by someone else",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "More than 30 requests per 60 s from the same IP",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Register a view of a video",
        "tags": [
          "videos"
        ]
      }
    },
    "/videos/{publicId}/reaction": {
      "put": {
        "description": "Records or switches the reaction of the authenticated user on the video and returns the like count after the operation. One reaction per user per video; repeating the same reaction changes nothing. There is no public dislike count. Rate limited to 60 requests per 60 s per IP.",
        "operationId": "VideosController_setReaction",
        "parameters": [
          {
            "name": "publicId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetReactionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Reaction recorded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReactionStateResponse"
                }
              }
            }
          },
          "400": {
            "description": "type missing or not one of like/dislike",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Video not found, or a draft of another channel",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "More than 60 requests per 60 s from the same IP",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Like or dislike a video",
        "tags": [
          "videos"
        ]
      },
      "delete": {
        "description": "Removes the reaction of the authenticated user and returns the like count after the operation. Idempotent: without a reaction, returns the current state. Rate limited to 60 requests per 60 s per IP.",
        "operationId": "VideosController_removeReaction",
        "parameters": [
          {
            "name": "publicId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Reaction removed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReactionStateResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Video not found, or a draft of another channel",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "More than 60 requests per 60 s from the same IP",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Remove the reaction from a video",
        "tags": [
          "videos"
        ]
      }
    },
    "/videos/{publicId}/suggestions": {
      "get": {
        "description": "Returns published, public videos from the same category as the reference video, newest first, excluding the reference video itself, drafts and unlisted videos. Paginated with offset/limit; the default page is 4. An empty list is a legitimate result, not an error.",
        "operationId": "VideosController_getSuggestions",
        "parameters": [
          {
            "name": "publicId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "description": "Quantas sugestões pular.",
            "schema": {
              "minimum": 0,
              "default": 0,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Quantas sugestões retornar.",
            "schema": {
              "minimum": 1,
              "default": 4,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Suggestions page",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuggestedVideosPage"
                }
              }
            }
          },
          "400": {
            "description": "offset or limit outside the accepted range",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Reference video not found, or a draft requested by someone else",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "List suggestions for the sidebar of the watch page",
        "tags": [
          "videos"
        ]
      }
    },
    "/videos/{publicId}/stream": {
      "get": {
        "description": "Redirects to a presigned storage URL for the video file. Accessible without authentication. A valid bearer token is optional and only identifies the channel owner, who may also stream their own drafts.",
        "operationId": "VideosController_stream",
        "parameters": [
          {
            "name": "publicId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "302": {
            "description": "Redirects to a presigned storage URL"
          },
          "404": {
            "description": "Video not found, or a draft requested by someone else",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "Video is not ready for streaming",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Stream a video",
        "tags": [
          "videos"
        ]
      }
    },
...
  },
  "info": {
    "title": "StreamTube API",
    "description": "API REST do StreamTube",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "ApiErrorEnvelope": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 401
          },
          "error": {
            "type": "string",
            "example": "INVALID_CREDENTIALS"
          },
          "message": {
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            ],
            "example": "Invalid email or password"
          },
          "code": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "statusCode",
          "error",
          "message"
        ]
      },
...
      "VideoDetailResponse": {
        "type": "object",
        "properties": {
          "publicId": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "enum": [
              "draft",
              "processing",
              "ready",
              "failed"
            ],
            "type": "string"
          },
          "durationSeconds": {
            "type": "number",
            "nullable": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "category": {
            "enum": [
              "Música",
              "Jogos",
              "Educação",
              "Entretenimento",
              "Notícias",
              "Esportes",
              "Tecnologia",
              "Outros"
            ],
            "type": "string"
          },
          "visibility": {
            "enum": [
              "public",
              "unlisted"
            ],
            "type": "string"
          },
          "publishedAt": {
            "type": "string",
            "nullable": true,
            "format": "date-time",
            "description": "Nulo quando o vídeo ainda é rascunho."
          },
          "thumbnailUrl": {
            "type": "string",
            "nullable": true,
            "description": "URL única já resolvida: a thumbnail customizada quando existe, senão a gerada pelo worker."
          }
        },
        "required": [
          "publicId",
          "title",
          "description",
          "status",
          "durationSeconds",
          "createdAt",
          "category",
          "visibility",
          "publishedAt",
          "thumbnailUrl"
        ]
      },
...
    }
  }
}